Пишет тесты для новых компонентов и хоков
