+++ src/components/films-list/films-list.test.js
@@ -8,6 +8,26 @@ import renderer from "react-test-renderer";
 
 const mockStore = configureStore([]);
 
+const film =  {
+  backgroundColor: `#444444`,
+  backgroundImage: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
+  description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
+  director: `David Yates`,
+  genre: `Fantasy`,
+  id: 1,
+  isFavourite: true,
+  image: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
+  preview: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
+  previewVideoLink: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
+  ratingScore: 7.2,
+  ratingCount: 248,
+  runTime: 223,
+  starring: [`Eddie Redmayne`, `Katherine Waterson`, `Dan Folger`],
+  title: `Fantastic Beasts: The Crimes of Grindelwald`,
+  videoLink: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
+  year: 2015,
+};
+
 const films = [
   {
     backgroundColor: `#444444`,
@@ -15,7 +35,7 @@ const films = [
     description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
     director: `David Yates`,
     genre: `Fantasy`,
-    id: 3523525245235,
+    id: 1,
     isFavourite: true,
     image: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
     preview: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
@@ -34,7 +54,7 @@ const films = [
     description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
     director: `Brayan Singer`,
     genre: `Story`,
-    id: Math.random(),
+    id: 2,
     isFavourite: true,
     image: `img/bohemian-rhapsody.jpg`,
     preview: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
@@ -54,7 +74,7 @@ const films = [
     description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
     director: `Brayan Singer`,
     genre: `Biography`,
-    id: Math.random(),
+    id: 3,
     isFavourite: true,
     image: `img/macbeth.jpg`,
     preview: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
@@ -74,7 +94,7 @@ const films = [
     description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
     director: `Martin Scoresese`,
     genre: `Biography`,
-    id: Math.random(),
+    id: 4,
     isFavourite: true,
     image: `img/aviator.jpg`,
     preview: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
@@ -117,7 +137,7 @@ const filmsList = [
     description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
     director: `David Yates`,
     genre: `Fantasy`,
-    id: 3523525245235,
+    id: 1,
     isFavourite: true,
     image: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
     preview: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
@@ -136,7 +156,7 @@ const filmsList = [
     description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
     director: `Brayan Singer`,
     genre: `Story`,
-    id: Math.random(),
+    id: 4,
     isFavourite: true,
     image: `img/bohemian-rhapsody.jpg`,
     preview: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
@@ -172,31 +192,11 @@ const filmsList = [
   },
 ];
 
-const film = {
-  backgroundColor: `#444444`,
-  backgroundImage: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
-  description: `In the 1930s, the Grand Budapest Hotel is a popular European ski resort, presided over by concierge Gustave H. (Ralph Fiennes). Zero, a junior lobby boy, becomes Gustave's friend and protege.`,
-  director: `David Yates`,
-  genre: `Fantasy`,
-  id: 3523525245235,
-  isFavourite: true,
-  image: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
-  preview: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
-  previewVideoLink: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-  ratingScore: 7.2,
-  ratingCount: 248,
-  runTime: 223,
-  starring: [`Eddie Redmayne`, `Katherine Waterson`, `Dan Folger`],
-  title: `Fantastic Beasts: The Crimes of Grindelwald`,
-  videoLink: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
-  year: 2015,
-};
-
 it(`FilmsList is rendered correctly`, () => {
   const store = mockStore({
     [NameSpace.DATA]: {
       allFilms: films,
-      film: film,
+      id: 1,
     },
   });
 
@@ -205,6 +205,14 @@ it(`FilmsList is rendered correctly`, () => {
         <Provider store={store}>
           <FilmsList
             films={filmsList}
+            id={1}
+            match={
+              {
+                  params: {
+                      id: 1
+                  }
+              }
+          }
           />
         </Provider>
       </BrowserRouter>, {
