+++ src/hocs/with-active-item/with-active-item.test.js
@@ -1,0 +1,27 @@
+import React from "react";
+import renderer from "react-test-renderer";
+import withActiveItem from "./with-active-item.js";
+
+const MockComponent = () => {
+  return (
+    <ul>
+    </ul>
+  );
+};
+
+const MockComponentWrapped = withActiveItem(MockComponent);
+
+it(`withActiveItem is rendered correctly`, () => {
+  const tree = renderer.create((
+    <MockComponentWrapped
+      activeItem={`active item`}
+      onActiveClick={() => {}}
+    />
+  ), {
+    createNodeMock() {
+      return {};
+    }
+  }).toJSON();
+
+  expect(tree).toMatchSnapshot();
+});
