+++ src/reducer/data/data.test.js
@@ -95,7 +95,7 @@ const films = [
 
 it(`Reducer without additional parameters should return initial state`, () => {
   expect(reducer(void 0, {})).toEqual({
-    allFilms: [],
+    allFilms: null,
     genre: Genre.ALL,
     favouriteFilms: [],
     promoFilm: {},
